Updates to IngredientMeasure, AddRecipe
I noticed that we declared the amount as an int; however, it's totally normal to have 0.25 tsp salt or 0.5 cup sugar. I mean, it's usually written 1/4 tsp and 1/2 cup. I changed int to Number. Thoughts?

This did break a thing or two in DatabaseHandler.

Also, everything seems to work with adding recipe directions into the form. There's a method that grabs these instructions and sends them to the RecipeBuilder.